requirements.txt
sqlalchemy[asyncio]==2.0.29
alembic==1.13.1
asyncpg==0.29.0
fastapi==0.110.3
pydantic==2.7.1
python-dotenv==1.0.1
uvicorn==0.29.0
pydantic_settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]
pyjwt
python-multipart